feat(ui): add welcome message entity
